refactor: separated initialization feature into hook
